ocapi/requirements.txt
boto3==1.24.13
coralogix_logger==2.0.5
